+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/model/Resource.java
@@ -119,16 +119,25 @@ public abstract class Resource extends BaseResource implements IAnyResource {
       return this;
     }
 
-    /**
-     * @return The logical id of the resource, as used in the URL for the resource. Once assigned, this value never changes.
-     */
+  /**
+   * @return The most complete id value of the resource, containing all
+   * available context and history. Once assigned this value never changes.
+   * NOTE: this value is NOT limited to just the logical id property of a
+   * resource id.
+   * @see IdType
+   * @see IdType#getValue()
+   */
     public String getId() { 
       return this.id == null ? null : this.id.getValue();
     }
 
-    /**
-     * @param value The logical id of the resource, as used in the URL for the resource. Once assigned, this value never changes.
-     */
+  /**
+   * @param value The id value of the resource. Once assigned, this value
+   *              never changes.
+   *
+   * @see IdType
+   * @see IdType#setValue(String)
+   */
     public Resource setId(String value) { 
       if (Utilities.noString(value))
         this.id = null;
@@ -139,6 +148,14 @@ public abstract class Resource extends BaseResource implements IAnyResource {
       }
       return this;
     }
+
+  /**
+   * @return  the logical ID part of this resource's id
+   * @see IdType#getIdPart()
+   */
+  public String getIdPart() {
+    return getIdElement().getIdPart();
+  }
 
     /**
      * @return {@link #meta} (The metadata about the resource. This is content that is maintained by the infrastructure. Changes to the content might not always be associated with version changes to the resource.)
@@ -414,8 +431,8 @@ public abstract class Resource extends BaseResource implements IAnyResource {
           , language);
       }
 
-// Manual code (from Configuration.txt):
-@Override
+  // Manual code (from Configuration.txt):
+  @Override
   public String getIdBase() {
     return getId();
   }