+++ RELEASE_NOTES.md
@@ -4,4 +4,5 @@
 
 ## Other code changes
 
-* no changes
+* Fix reverse references in GraphQL searches
+